handson-app/requirements.txt
azure-cosmos
azure-functions
openai==1.*
